Moved SnackbarUtils to the androidshared module
