+++ androidshared/src/main/java/org/odk/collect/androidshared/ui/SnackbarUtils.kt
@@ -11,7 +11,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package org.odk.collect.android.utilities
+package org.odk.collect.androidshared.ui
 
 import android.view.View
 import android.widget.TextView
